64-bit indexing: replace uint/int with size_t/ptrdiff_t

+++ src/template/encode1.c
@@ -2,7 +2,7 @@
 
 /* gather 4-value block from strided array */
 static void
-_t2(gather, Scalar, 1)(Scalar* q, const Scalar* p, int sx)
+_t2(gather, Scalar, 1)(Scalar* q, const Scalar* p, ptrdiff_t sx)
 {
   uint x;
   for (x = 0; x < 4; x++, p += sx)
@@ -11,9 +11,9 @@ _t2(gather, Scalar, 1)(Scalar* q, const Scalar* p, int sx)
 
 /* gather nx-value block from strided array */
 static void
-_t2(gather_partial, Scalar, 1)(Scalar* q, const Scalar* p, uint nx, int sx)
+_t2(gather_partial, Scalar, 1)(Scalar* q, const Scalar* p, size_t nx, ptrdiff_t sx)
 {
-  uint x;
+  size_t x;
   for (x = 0; x < nx; x++, p += sx)
     q[x] = *p;
   _t1(pad_block, Scalar)(q, nx, 1);
@@ -30,8 +30,8 @@ _t2(fwd_xform, Int, 1)(Int* p)
 /* public functions -------------------------------------------------------- */
 
 /* encode 4-value block stored at p using stride sx */
-uint
-_t2(zfp_encode_block_strided, Scalar, 1)(zfp_stream* stream, const Scalar* p, int sx)
+size_t
+_t2(zfp_encode_block_strided, Scalar, 1)(zfp_stream* stream, const Scalar* p, ptrdiff_t sx)
 {
   /* gather block from strided array */
   cache_align_(Scalar block[4]);
@@ -41,8 +41,8 @@ _t2(zfp_encode_block_strided, Scalar, 1)(zfp_stream* stream, const Scalar* p, in
 }
 
 /* encode nx-value block stored at p using stride sx */
-uint
-_t2(zfp_encode_partial_block_strided, Scalar, 1)(zfp_stream* stream, const Scalar* p, uint nx, int sx)
+size_t
+_t2(zfp_encode_partial_block_strided, Scalar, 1)(zfp_stream* stream, const Scalar* p, size_t nx, ptrdiff_t sx)
 {
   /* gather block from strided array */
   cache_align_(Scalar block[4]);
