+++ tests/src/endtoend/serialExecBase.c
@@ -180,7 +180,7 @@ isCompressedValuesWithinAccuracy(struct setupVars* bundle)
     return 1;
   }
 
-  int strides[4];
+  ptrdiff_t strides[4];
   zfp_field_stride(field, strides);
 
   // apply strides
